+++ src/XrmDefinitelyTyped/AssemblyInfo.fs
@@ -6,9 +6,9 @@ open System.Reflection
 [<assembly: AssemblyDescriptionAttribute("Tool to generate TypeScript declaration files for MS CRM Dynamics client-side coding.")>]
 [<assembly: AssemblyCompanyAttribute("Delegate A/S")>]
 [<assembly: AssemblyCopyrightAttribute("Copyright (c) Delegate A/S 2015")>]
-[<assembly: AssemblyVersionAttribute("1.10.0")>]
-[<assembly: AssemblyFileVersionAttribute("1.10.0")>]
+[<assembly: AssemblyVersionAttribute("1.11.0")>]
+[<assembly: AssemblyFileVersionAttribute("1.11.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.10.0"
+    let [<Literal>] Version = "1.11.0"