+++ src/XrmDefinitelyTyped/Resources/dg.xrmquery.d.ts
@@ -50,8 +50,26 @@ declare module XQC {
         orderDesc(vars: (x: S) => Attribute<S>): this;
         skip(amount: number): this;
         top(amount: number): this;
-        execute(successCallback: (records: R[]) => any, errorCallback?: (err: Error) => any, onComplete?: () => any): void;
-        getFirst(successCallback: (record: R) => any, errorCallback?: (err: Error) => any): void;
+        /**
+         * Executes the RetrieveMultiple. Note that the first function passed as an argument is called once per page returned from CRM.
+         * @param pageSuccessCallback Called once per page returned from CRM
+         * @param errorCallback Called if an error occurs during the retrieval
+         * @param onComplete Called when all pages have been successfully retrieved from CRM
+         */
+        execute(pageSuccessCallback: (records: R[]) => any, errorCallback: (err: Error) => any, onComplete: () => any): void;
+        /**
+         * Executes the RetrieveMultiple and concatenates all the pages to a single array that is delivered to the success callback function.
+         * @param successCallback Called with all records returned from the query
+         * @param errorCallback Called if an error occures during the retrieval
+         */
+        getAll(successCallback: (records: R[]) => any, errorCallback?: (err: Error) => any): void;
+        /**
+         * Executes the RetrieveMultiple, but only returns the first result (or null, if no record was found).
+         * @param successCallback Called with the first result of the query (or null, if no record was found)
+         * @param errorCallback Called if an error occures during the retrieval
+         */
+        getFirst(successCallback: (record: R | null) => any, errorCallback?: (err: Error) => any): void;
+        getOptionString(): string;
     }
     /**
      * Contains information about a Create query